diario/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
bcrypt==4.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==9.1.1
Flask-Bcrypt==1.0.1